refactor(-): :hammer: move models solicitacao dieta
