+++ sme_sigpae_api/dados_comuns/__tests__/conftest.py
@@ -22,7 +22,7 @@ from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadas,
     LogQuantidadeDietasAutorizadasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     ClassificacaoDieta,
 )
 from sme_sigpae_api.perfil.models import ContentType
