+++ sme_sigpae_api/dados_comuns/__tests__/test_metodos_carga_dados.py
@@ -11,12 +11,10 @@ from sme_sigpae_api.cardapio.base.models import (
 )
 from sme_sigpae_api.cardapio.suspensao_alimentacao.models import MotivoSuspensao
 from sme_sigpae_api.dados_comuns.models import Contato, TemplateMensagem
-from sme_sigpae_api.dieta_especial.models import (
-    AlergiaIntolerancia,
-    Alimento,
-    ClassificacaoDieta,
-)
+from sme_sigpae_api.dieta_especial.models import Alimento
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    AlergiaIntolerancia,
+    ClassificacaoDieta,
     MotivoAlteracaoUE,
     MotivoNegacao,
 )
