+++ sme_sigpae_api/dieta_especial/fixtures/factories/dieta_especial_base_factory.py
@@ -3,14 +3,14 @@ from factory.django import DjangoModelFactory
 from faker import Faker
 
 from sme_sigpae_api.dieta_especial.models import (
-    AlergiaIntolerancia,
     Alimento,
-    ClassificacaoDieta,
     ProtocoloPadraoDietaEspecial,
     SolicitacaoDietaEspecial,
     SubstituicaoAlimento,
 )
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    AlergiaIntolerancia,
+    ClassificacaoDieta,
     MotivoAlteracaoUE,
 )
 from sme_sigpae_api.escola.fixtures.factories.escola_factory import (
