+++ sme_sigpae_api/dieta_especial/management/commands/dieta_especial_autorizar.py
@@ -2,11 +2,13 @@ from django.core.management.base import BaseCommand
 from faker import Faker
 
 from sme_sigpae_api.dieta_especial.models import (
-    AlergiaIntolerancia,
     Alimento,
-    ClassificacaoDieta,
     SolicitacaoDietaEspecial,
     SubstituicaoAlimento,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    AlergiaIntolerancia,
+    ClassificacaoDieta,
 )
 from sme_sigpae_api.perfil.models import Usuario
