+++ sme_sigpae_api/dieta_especial/solicitacao_dieta_especial/models.py
@@ -63,6 +63,42 @@ CANCELADOS_EVENTO_DIETA_ESPECIAL_TEMP = [
 ]
 
 
+class MotivoNegacao(Descritivel):
+    CANCELAMENTO = "CANCELAMENTO"
+    INCLUSAO = "INCLUSAO"
+
+    PROCESSO_CHOICES = (
+        (CANCELAMENTO, "Solicitação de Cancelamento"),
+        (INCLUSAO, "Solicitação de Inclusão"),
+    )
+
+    processo = models.CharField(
+        choices=PROCESSO_CHOICES, default=INCLUSAO, blank=False, max_length=20
+    )
+
+    def __str__(self):
+        return self.descricao
+
+
+class MotivoAlteracaoUE(Descritivel, Nomeavel, TemChaveExterna, Ativavel):
+    def __str__(self):
+        return self.nome
+
+    class Meta:
+        verbose_name = "Motivo Alteração U.E"
+        verbose_name_plural = "Motivo Alteração U.E"
+
+
+class AlergiaIntolerancia(Descritivel):
+    def __str__(self):
+        return self.descricao
+
+
+class ClassificacaoDieta(Descritivel, Nomeavel):
+    def __str__(self):
+        return self.nome
+
+
 class SolicitacaoDietaEspecial(
     ExportModelOperationsMixin("dieta_especial"),
     TemChaveExterna,
@@ -426,29 +462,3 @@ class Anexo(ExportModelOperationsMixin("anexo"), models.Model):
 
     def __str__(self):
         return self.nome
-
-
-class MotivoNegacao(Descritivel):
-    CANCELAMENTO = "CANCELAMENTO"
-    INCLUSAO = "INCLUSAO"
-
-    PROCESSO_CHOICES = (
-        (CANCELAMENTO, "Solicitação de Cancelamento"),
-        (INCLUSAO, "Solicitação de Inclusão"),
-    )
-
-    processo = models.CharField(
-        choices=PROCESSO_CHOICES, default=INCLUSAO, blank=False, max_length=20
-    )
-
-    def __str__(self):
-        return self.descricao
-
-
-class MotivoAlteracaoUE(Descritivel, Nomeavel, TemChaveExterna, Ativavel):
-    def __str__(self):
-        return self.nome
-
-    class Meta:
-        verbose_name = "Motivo Alteração U.E"
-        verbose_name_plural = "Motivo Alteração U.E"
