+++ sme_sigpae_api/dieta_especial/tasks/utils/logs.py
@@ -13,7 +13,7 @@ from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadasRecreioNasFerias,
     LogQuantidadeDietasAutorizadasRecreioNasFeriasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     ClassificacaoDieta,
 )
 from sme_sigpae_api.escola.models import FaixaEtaria, PeriodoEscolar
