+++ sme_sigpae_api/medicao_inicial/__tests__/test_serializers.py
@@ -8,7 +8,7 @@ from rest_framework import serializers
 from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     ClassificacaoDieta,
 )
 from sme_sigpae_api.escola.fixtures.factories.escola_factory import (
