+++ sme_sigpae_api/paineis_consolidados/utils/totalizadores_relatorio.py
@@ -1,9 +1,9 @@
 from collections import Counter
 
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     AlergiaIntolerancia,
     ClassificacaoDieta,
-    SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.escola.models import Escola, Lote, TipoGestao, TipoUnidadeEscolar
