+++ utility/carga_dados/medicao/insere_informacoes_lancamento_inicial.py
@@ -19,7 +19,7 @@ from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadas,
     LogQuantidadeDietasAutorizadasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     ClassificacaoDieta,
 )
 from sme_sigpae_api.escola.models import (
